Create score keeping
Added potential code to keep score & present the final score.
Added colors, but not sure if the format is correct. 
This calculates the score, but I am not sure if it will display it correctly. Next  step is to figure  out how to display the score periodically so the user can see it at all times.
2 hours